requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
djangorestframework==3.15.2
inject==5.2.1
marshmallow==3.23.1
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2